requirements.txt
streamlit
pytesseract
pillow
streamlit-copy-to-clipboard